+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `name` option to `write_avro` method
 - Added support for array of name-type pairs to `schema` option
 - Added `cast` method to `DataFrame` and `LazyFrame`
+- Fixed `read_database` connection leasing for Active Record 7.2
 
 ## 0.14.0 (2024-09-17)